requirements.txt
google-adk
python-dotenv
litellm
duckduckgo_search
ddgs
mcp